fix(ci): green lint/test/hassfest + runtime service & dashboard bugs
Lint/format/tests:
- ruff: fix 54 findings (imports, datetime.UTC, ClassVar on mutable class
  defaults, contextlib.suppress, combined with, dict comprehension, drop
  unused import); apply ruff format across the tree.
- tests: add tests/__init__.py so `from .conftest import ...` resolves
  (collection was failing); fix wrong blade-runtime entity_id in test.

hassfest:
- services.yaml: replace forbidden `target: device` filters with entity
  selectors on all 10 services (handler already resolves entity->entry).
- manifest.json: reorder keys (domain, name, then alphabetical).

Runtime fixes:
- services: handlers were registered as sync lambdas returning un-awaited
  coroutines, so every wrapped service silently no-op'd at runtime. Register
  genuine coroutine functions instead.
- lawn_mower: force immediate async_refresh() after commands so rapid
  start->pause sequences reflect at once (async_request_refresh debounced).
- config_flow: gate aiohttp session creation on a new backend.requires_session
  flag so skeleton backends don't spawn an unused DNS resolver thread.

Dashboard/docs:
- fix broken entity refs: sensor.tron_blade_runtime ->
  sensor.tron_blade_runtime_hours and _blade_life_remaining ->
  _blade_life_remaining_percent (match deployed entity_ids).

+++ custom_components/airseekers/device_tracker.py
@@ -46,4 +46,4 @@ class AirseekersDeviceTracker(AirseekersEntity, TrackerEntity):
     @property
     def location_accuracy(self) -> int:
         acc = self._status.gps_accuracy
-        return int(round(acc)) if acc is not None else 0
+        return round(acc) if acc is not None else 0
